SCUMM: Move Mac STRS handling to macgui_strings.cpp
